requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
google-genai>=0.2.2
gspread>=6.0.0
google-auth>=2.27.0
pytz>=2023.3.post1
pytest>=8.0.2
